Initial commit.
This is basically an extraction and refactoring of an error report
handling code from my tetris clone project.

I'm turning it into a standalone library so that I can drop it into
other C projects of mine.

The code here isn't tested. It came from another project which is
working, but the code here was slightly modified, which means bugs
may have been introduced.